backend/requirements.txt
fastapi
uvicorn
firebase-admin
python-multipart
pydantic
python-dotenv
pypdf
python-docx
openpyxl
websockets
python-pptx
beautifulsoup4
lxml
requests
psycopg2-binary
asyncpg
redis
